RUMM-500 Organize mocks by feature
