+++ Sources/Datadog/Core/Attributes/UserInfo.swift
