--- Tests/DatadogTests/Datadog/Mocks/DatadogTracingMocks.swift
@@ -1,78 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2019-2020 Datadog, Inc.
- */
-
-import Foundation
-import XCTest
-@testable import Datadog
-
-/*
-A collection of SDK object mocks for Tracing.
-It follows the mocking conventions described in `FoundationMocks.swift`.
- */
-
-class RelativeTracingUUIDGenerator: TracingUUIDGenerator {
-    private(set) var uuid: TracingUUID
-    internal let count: UInt64
-    private let queue = DispatchQueue(label: "queue-RelativeTracingUUIDGenerator-\(UUID().uuidString)")
-
-    init(startingFrom uuid: TracingUUID, advancingByCount count: UInt64 = 1) {
-        self.uuid = uuid
-        self.count = count
-    }
-
-    func generateUnique() -> TracingUUID {
-        return queue.sync {
-            defer { uuid = TracingUUID(rawValue: uuid.rawValue + count) }
-            return uuid
-        }
-    }
-}
-
-extension TracingUUID {
-    static func mockAny() -> TracingUUID {
-        return TracingUUID(rawValue: .mockAny())
-    }
-
-    static func mock(_ rawValue: UInt64) -> TracingUUID {
-        return TracingUUID(rawValue: rawValue)
-    }
-}
-
-// MARK: - Integration
-
-/// `SpanOutput` recording received spans.
-class SpanOutputMock: SpanOutput {
-    struct Recorded {
-        let span: DDSpan
-        let finishTime: Date
-    }
-
-    var recorded: Recorded? = nil
-
-    func write(ddspan: DDSpan, finishTime: Date) {
-        recorded = Recorded(span: ddspan, finishTime: finishTime)
-    }
-}
-
-extension DDTracer {
-    static func mockAny() -> DDTracer {
-        return mockWith()
-    }
-
-    static func mockWith(
-        spanOutput: SpanOutput = SpanOutputMock(),
-        logOutput: LoggingForTracingAdapter.AdaptedLogOutput = .init(loggingOutput: LogOutputMock()),
-        dateProvider: DateProvider = SystemDateProvider(),
-        tracingUUIDGenerator: TracingUUIDGenerator = DefaultTracingUUIDGenerator()
-    ) -> DDTracer {
-        return DDTracer(
-            spanOutput: spanOutput,
-            logOutput: logOutput,
-            dateProvider: dateProvider,
-            tracingUUIDGenerator: tracingUUIDGenerator
-        )
-    }
-}
