--- Tests/DatadogTests/Datadog/Mocks/LogsMocks.swift
@@ -1,87 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2019-2020 Datadog, Inc.
- */
-
-import Foundation
-@testable import Datadog
-
-/*
-A collection of mocks for `Log` objects.
-It follows the mocking conventions described in `FoundationMocks.swift`.
- */
-
-extension Log {
-    static func mockWith(
-        date: Date = .mockAny(),
-        status: Log.Status = .mockAny(),
-        message: String = .mockAny(),
-        serviceName: String = .mockAny(),
-        environment: String = .mockAny(),
-        loggerName: String = .mockAny(),
-        loggerVersion: String = .mockAny(),
-        threadName: String = .mockAny(),
-        applicationVersion: String = .mockAny(),
-        userInfo: UserInfo = .mockAny(),
-        networkConnectionInfo: NetworkConnectionInfo = .mockAny(),
-        mobileCarrierInfo: CarrierInfo? = .mockAny(),
-        attributes: [String: EncodableValue]? = nil,
-        tags: [String]? = nil
-    ) -> Log {
-        return Log(
-            date: date,
-            status: status,
-            message: message,
-            serviceName: serviceName,
-            environment: environment,
-            loggerName: loggerName,
-            loggerVersion: loggerVersion,
-            threadName: threadName,
-            applicationVersion: applicationVersion,
-            userInfo: userInfo,
-            networkConnectionInfo: networkConnectionInfo,
-            mobileCarrierInfo: mobileCarrierInfo,
-            attributes: attributes,
-            tags: tags
-        )
-    }
-}
-
-extension Log.Status {
-    static func mockAny() -> Log.Status {
-        return .info
-    }
-}
-
-extension EncodableValue {
-    static func mockAny() -> EncodableValue {
-        return EncodableValue(String.mockAny())
-    }
-}
-
-extension LogBuilder {
-    static func mockAny() -> LogBuilder {
-        return mockWith()
-    }
-
-    static func mockWith(
-        applicationVersion: String = .mockAny(),
-        environment: String = .mockAny(),
-        serviceName: String = .mockAny(),
-        loggerName: String = .mockAny(),
-        userInfoProvider: UserInfoProvider = .mockAny(),
-        networkConnectionInfoProvider: NetworkConnectionInfoProviderType = NetworkConnectionInfoProviderMock.mockAny(),
-        carrierInfoProvider: CarrierInfoProviderType = CarrierInfoProviderMock.mockAny()
-    ) -> LogBuilder {
-        return LogBuilder(
-            applicationVersion: applicationVersion,
-            environment: environment,
-            serviceName: serviceName,
-            loggerName: loggerName,
-            userInfoProvider: userInfoProvider,
-            networkConnectionInfoProvider: networkConnectionInfoProvider,
-            carrierInfoProvider: carrierInfoProvider
-        )
-    }
-}
