--- Tests/DatadogTests/Datadog/Mocks/SpansMocks.swift
@@ -1,77 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2019-2020 Datadog, Inc.
- */
-
-import Foundation
-@testable import Datadog
-
-/*
-A collection of mocks for `DDSpan` objects.
-It follows the mocking conventions described in `FoundationMocks.swift`.
- */
-
-extension DDSpanContext {
-    static func mockAny() -> DDSpanContext {
-        return mockWith()
-    }
-
-    static func mockWith(
-        traceID: TracingUUID = .mockAny(),
-        spanID: TracingUUID = .mockAny(),
-        parentSpanID: TracingUUID? = .mockAny()
-    ) -> DDSpanContext {
-        return DDSpanContext(
-            traceID: traceID,
-            spanID: spanID,
-            parentSpanID: parentSpanID
-        )
-    }
-}
-
-extension DDSpan {
-    static func mockAny() -> DDSpan {
-        return mockWith()
-    }
-
-    static func mockWith(
-        tracer: DDTracer = .mockAny(),
-        context: DDSpanContext = .mockAny(),
-        operationName: String = .mockAny(),
-        startTime: Date = .mockAny(),
-        tags: [String: Codable] = [:]
-    ) -> DDSpan {
-        return DDSpan(
-            tracer: tracer,
-            context: context,
-            operationName: operationName,
-            startTime: startTime,
-            tags: tags
-        )
-    }
-}
-
-extension SpanBuilder {
-    static func mockAny() -> SpanBuilder {
-        return mockWith()
-    }
-
-    static func mockWith(
-        applicationVersion: String = .mockAny(),
-        environment: String = .mockAny(),
-        serviceName: String = .mockAny(),
-        userInfoProvider: UserInfoProvider = .mockAny(),
-        networkConnectionInfoProvider: NetworkConnectionInfoProviderType = NetworkConnectionInfoProviderMock.mockAny(),
-        carrierInfoProvider: CarrierInfoProviderType = CarrierInfoProviderMock.mockAny()
-    ) -> SpanBuilder {
-        return SpanBuilder(
-            applicationVersion: applicationVersion,
-            environment: environment,
-            serviceName: serviceName,
-            userInfoProvider: userInfoProvider,
-            networkConnectionInfoProvider: networkConnectionInfoProvider,
-            carrierInfoProvider: carrierInfoProvider
-        )
-    }
-}
